CloudPubSubFun/requirements.txt
pandas
tweepy
google-cloud-bigquery
google-cloud-pubsub